acbc_app/requirements.txt
asgiref==3.8.0
boto3==1.34.131
botocore==1.34.131
certifi==2024.6.2
chardet==5.2.0
dj-database-url==2.2.0
Django==5.0
django-allauth[socialaccount]==64.0.0
PyJWT==2.8.0
cryptography==43.0.0
django-braces==1.15.0
django-cors-headers==4.4.0
django-model-utils==4.5.1
django-notifications-hq==1.8.3
django-settings-export==1.2.1
django-storages==1.14.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-rest-auth[with_social]==7.0.1
factory-boy==3.3.0
gunicorn==22.0.0
idna==3.7
jmespath==1.0.1
packaging==24.1
Pillow==10.3.0
psycopg[binary]==3.1.19
pyparsing==3.1.2
python-dateutil==2.8.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.1
setuptools>=69.0.3
six==1.16.0
sqlparse==0.5.0
swapper==1.3.0
urllib3==2.2.2
webencodings==0.5.1
whitenoise==6.7.0
pymupdf==1.24.2
bleach==6.1.0
drf-yasg==1.21.14
inflection==0.5.1
uritemplate==4.1.1
beautifulsoup4>=4.9.3
python-magic>=0.4.24
sentry-sdk>=2.0.0
postmarker>=1.0.0
